pyproject.toml
[project]
name = "orkarv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cocotb>=2.0.0",
    "cocotb-bus",
    "cocotbext-axi>=0.1.26",
    "cocotbext-jtag>=0.2.1",
    "cocotbext-uart>=0.1.4",
    "peakrdl>=1.5.0",
    "peakrdl-markdown>=1.0.3",
    "peakrdl-regblock-vhdl>=1.1.1.0",
    "peakrdl-regblock-vhdl-csr>=1.1.1.0",
    "peakrdl-rust[cli]>=0.4.0",
    "pydantic>=2.12.5",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.sources]
peakrdl-regblock-vhdl-csr = { path = "./submodules/PeakRDL-regblock-vhdl", editable = true }
